# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ErrorMessage {
  message: string;
}

export interface HealthStatus {
  status: string;
}

export interface CharacterInput {
  /** @minLength 1 */
  name: string;
  photoUri?: string;
  /** Optional short appearance description for illustration consistency. */
  appearance?: string;
}

export type StoryMode = (typeof StoryMode)[keyof typeof StoryMode];

export const StoryMode = {
  behavior: "behavior",
  random: "random",
} as const;

export interface SupportingCharacter {
  name: string;
  relationship: string;
}

export interface GenerateStoryRequest {
  mode: StoryMode;
  /** Parent-provided behavior focus or story idea. */
  prompt?: string;
  character: CharacterInput;
  /** Optional list of other characters to include if the story calls for them. */
  supportingCharacters?: SupportingCharacter[];
}

export interface StoryPage {
  /** @minimum 1 */
  pageNumber: number;
  /** 1-3 sentences of simple, age-appropriate story text. */
  text: string;
  /** A short description of what the watercolor illustration on this page should show. */
  illustrationPrompt: string;
  /** URL of the sliced storyboard image for this page. */
  imageUrl?: string;
  scene?: string;
  composition?: string;
  emotion?: string;
}

export type StoryGenerationStep =
  (typeof StoryGenerationStep)[keyof typeof StoryGenerationStep];

export const StoryGenerationStep = {
  queued: "queued",
  writing_story: "writing_story",
  painting_sheet: "painting_sheet",
  slicing_pages: "slicing_pages",
  preparing_reader: "preparing_reader",
  complete: "complete",
  failed: "failed",
} as const;

export type StoryGenerationStatusValue =
  (typeof StoryGenerationStatusValue)[keyof typeof StoryGenerationStatusValue];

export const StoryGenerationStatusValue = {
  queued: "queued",
  running: "running",
  complete: "complete",
  failed: "failed",
} as const;

export interface StoryGenerationJob {
  bookId: string;
  status: StoryGenerationStatusValue;
  step: StoryGenerationStep;
  message: string;
}

export type StoryGenerationStatus = StoryGenerationJob & {
  error?: string;
};

export interface StoryArtifactLinks {
  bookHtmlUrl?: string;
  storyJsonUrl?: string;
  usageJsonUrl?: string;
}

export interface GeneratedStory {
  title: string;
  /**
   * @minItems 10
   * @maxItems 20
   */
  pages: StoryPage[];
  reflectionQuestion: string;
  /** URL of the AI-generated watercolor cover illustration. */
  coverImageUrl?: string;
  /** URL of the sliced end-page illustration. */
  endImageUrl?: string;
  /** URL of the full 4x4 storyboard sheet. */
  sheetImageUrl?: string;
  artifactLinks?: StoryArtifactLinks;
}