# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stories
    description: Story generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stories/generate:
    post:
      operationId: generateStory
      tags: [stories]
      summary: Generate a child picture book
      description: Starts a high-quality story-sheet generation job for a child character.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStoryRequest"
      responses:
        "202":
          description: Story generation job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryGenerationJob"
        "503":
          description: Story generation is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /stories/{bookId}/status:
    get:
      operationId: getStoryStatus
      tags: [stories]
      summary: Get story generation status
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Story generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryGenerationStatus"
        "404":
          description: Story job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /stories/{bookId}:
    get:
      operationId: getGeneratedStory
      tags: [stories]
      summary: Get completed generated story
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated picture book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedStory"
        "404":
          description: Story job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "409":
          description: Story generation is not complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryGenerationStatus"
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CharacterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        photoUri:
          type: string
        appearance:
          type: string
          description: Optional short appearance description for illustration consistency.
      required:
        - name
    StoryMode:
      type: string
      enum:
        - behavior
        - random
    SupportingCharacter:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
      required:
        - name
        - relationship
    GenerateStoryRequest:
      type: object
      properties:
        mode:
          $ref: "#/components/schemas/StoryMode"
        prompt:
          type: string
          description: Parent-provided behavior focus or story idea.
        character:
          $ref: "#/components/schemas/CharacterInput"
        supportingCharacters:
          type: array
          items:
            $ref: "#/components/schemas/SupportingCharacter"
          description: Optional list of other characters to include if the story calls for them.
      required:
        - mode
        - character
    StoryPage:
      type: object
      properties:
        pageNumber:
          type: integer
          minimum: 1
        text:
          type: string
          description: 1-3 sentences of simple, age-appropriate story text.
        illustrationPrompt:
          type: string
          description: A short description of what the watercolor illustration on this page should show.
        imageUrl:
          type: string
          description: URL of the sliced storyboard image for this page.
        scene:
          type: string
        composition:
          type: string
        emotion:
          type: string
      required:
        - pageNumber
        - text
        - illustrationPrompt
    StoryGenerationStep:
      type: string
      enum:
        - queued
        - writing_story
        - painting_sheet
        - slicing_pages
        - preparing_reader
        - complete
        - failed
    StoryGenerationStatusValue:
      type: string
      enum:
        - queued
        - running
        - complete
        - failed
    StoryGenerationJob:
      type: object
      properties:
        bookId:
          type: string
        status:
          $ref: "#/components/schemas/StoryGenerationStatusValue"
        step:
          $ref: "#/components/schemas/StoryGenerationStep"
        message:
          type: string
      required:
        - bookId
        - status
        - step
        - message
    StoryGenerationStatus:
      allOf:
        - $ref: "#/components/schemas/StoryGenerationJob"
        - type: object
          properties:
            error:
              type: string
    StoryArtifactLinks:
      type: object
      properties:
        bookHtmlUrl:
          type: string
        storyJsonUrl:
          type: string
        usageJsonUrl:
          type: string
    GeneratedStory:
      type: object
      properties:
        title:
          type: string
        pages:
          type: array
          items:
            $ref: "#/components/schemas/StoryPage"
          minItems: 10
          maxItems: 20
        reflectionQuestion:
          type: string
        coverImageUrl:
          type: string
          description: URL of the AI-generated watercolor cover illustration.
        endImageUrl:
          type: string
          description: URL of the sliced end-page illustration.
        sheetImageUrl:
          type: string
          description: URL of the full 4x4 storyboard sheet.
        artifactLinks:
          $ref: "#/components/schemas/StoryArtifactLinks"
      required:
        - title
        - pages
        - reflectionQuestion