# lib/api-zod/src/generated/types/generatedStory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryArtifactLinks } from "./storyArtifactLinks";
import type { StoryPage } from "./storyPage";

export interface GeneratedStory {
  title: string;
  /**
   * @minItems 10
   * @maxItems 20
   */
  pages: StoryPage[];
  reflectionQuestion: string;
  /** URL of the AI-generated watercolor cover illustration. */
  coverImageUrl?: string;
  /** URL of the sliced end-page illustration. */
  endImageUrl?: string;
  /** URL of the full 4x4 storyboard sheet. */
  sheetImageUrl?: string;
  artifactLinks?: StoryArtifactLinks;
}