# lib/api-zod/src/generated/types/storyGenerationJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryGenerationStatusValue } from "./storyGenerationStatusValue";
import type { StoryGenerationStep } from "./storyGenerationStep";

export interface StoryGenerationJob {
  bookId: string;
  status: StoryGenerationStatusValue;
  step: StoryGenerationStep;
  message: string;
}