# lib/api-zod/src/generated/types/storyGenerationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryGenerationJob } from "./storyGenerationJob";

export type StoryGenerationStatus = StoryGenerationJob & {
  error?: string;
};