# lib/api-zod/src/generated/types/storyGenerationStatusValue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type StoryGenerationStatusValue =
  (typeof StoryGenerationStatusValue)[keyof typeof StoryGenerationStatusValue];

export const StoryGenerationStatusValue = {
  queued: "queued",
  running: "running",
  complete: "complete",
  failed: "failed",
} as const;