# lib/api-zod/src/generated/types/storyGenerationStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type StoryGenerationStep =
  (typeof StoryGenerationStep)[keyof typeof StoryGenerationStep];

export const StoryGenerationStep = {
  queued: "queued",
  writing_story: "writing_story",
  painting_sheet: "painting_sheet",
  slicing_pages: "slicing_pages",
  preparing_reader: "preparing_reader",
  complete: "complete",
  failed: "failed",
} as const;