# lib/api-zod/src/generated/types/storyPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoryPage {
  /** @minimum 1 */
  pageNumber: number;
  /** 1-3 sentences of simple, age-appropriate story text. */
  text: string;
  /** A short description of what the watercolor illustration on this page should show. */
  illustrationPrompt: string;
  /** URL of the sliced storyboard image for this page. */
  imageUrl?: string;
  scene?: string;
  composition?: string;
  emotion?: string;
}